[Oztechan/CCC#1815] Move android only modules to android:core
