--- ad/build.gradle.kts
@@ -1,71 +1,0 @@
-import config.BuildType
-import config.DeviceFlavour
-import config.Keys
-
-plugins {
-    @Suppress("DSL_SCOPE_VIOLATION")
-    libs.plugins.apply {
-        id(androidLib.get().pluginId)
-        id(android.get().pluginId)
-    }
-}
-
-@Suppress("UnstableApiUsage")
-android {
-    ProjectSettings.apply {
-        namespace = Modules.ad.packageName
-        compileSdk = COMPILE_SDK_VERSION
-
-        defaultConfig {
-            minSdk = MIN_SDK_VERSION
-            targetSdk = TARGET_SDK_VERSION
-        }
-    }
-
-    DeviceFlavour.apply {
-        flavorDimensions.addAll(listOf(flavorDimension))
-
-        productFlavors {
-            create(google) {
-                dimension = flavorDimension
-            }
-
-            create(huawei) {
-                dimension = flavorDimension
-            }
-        }
-    }
-
-    buildTypes {
-        getByName(BuildType.release) {
-            Keys(project, BuildType.RELEASE).apply {
-                resValue(typeString, admobAppId.resourceKey, admobAppId.value)
-                resValue(typeString, bannerAdIdCalculator.resourceKey, bannerAdIdCalculator.value)
-                resValue(typeString, bannerAdIdSettings.resourceKey, bannerAdIdSettings.value)
-                resValue(typeString, bannerAdIdCurrencies.resourceKey, bannerAdIdCurrencies.value)
-                resValue(typeString, interstitialAdId.resourceKey, interstitialAdId.value)
-                resValue(typeString, rewardedAdId.resourceKey, rewardedAdId.value)
-            }
-        }
-
-        getByName(BuildType.debug) {
-            Keys(project, BuildType.DEBUG).apply {
-                resValue(typeString, admobAppId.resourceKey, admobAppId.value)
-                resValue(typeString, bannerAdIdCalculator.resourceKey, bannerAdIdCalculator.value)
-                resValue(typeString, bannerAdIdSettings.resourceKey, bannerAdIdSettings.value)
-                resValue(typeString, bannerAdIdCurrencies.resourceKey, bannerAdIdCurrencies.value)
-                resValue(typeString, interstitialAdId.resourceKey, interstitialAdId.value)
-                resValue(typeString, rewardedAdId.resourceKey, rewardedAdId.value)
-            }
-        }
-    }
-}
-
-dependencies {
-    implementation(libs.common.koinCore)
-
-    @Suppress("UnstableApiUsage")
-    DeviceFlavour.googleImplementation(libs.android.google.admob)
-
-    implementation(project(Modules.Submodules.logmob))
-}
