--- ad/src/google/kotlin/com/oztechan/ccc/ad/Ads.kt
@@ -1,10 +1,0 @@
-package com.oztechan.ccc.ad
-
-import android.content.Context
-import co.touchlab.kermit.Logger
-import com.google.android.gms.ads.MobileAds
-
-fun initAds(context: Context) {
-    Logger.i { "Ads initAds" }
-    MobileAds.initialize(context)
-}
