--- ad/src/huawei/kotlin/com/oztechan/ccc/ad/Ads.kt
@@ -1,9 +1,0 @@
-package com.oztechan.ccc.ad
-
-import android.content.Context
-import co.touchlab.kermit.Logger
-
-@Suppress("UNUSED_PARAMETER")
-fun initAds(context: Context) {
-    Logger.i { "Ads initAds" }
-}
