--- ad/src/main/kotlin/com/oztechan/ccc/ad/BannerAdView.kt
@@ -1,15 +1,0 @@
-package com.oztechan.ccc.ad
-
-import android.content.Context
-import android.view.View
-import android.widget.FrameLayout
-
-class BannerAdView(
-    context: Context,
-    banner: View,
-    val onDestroy: () -> Unit
-) : FrameLayout(context) {
-    init {
-        this.addView(banner)
-    }
-}
