+++ android/app/src/main/kotlin/com/oztechan/ccc/android/app/Application.kt
@@ -12,9 +12,9 @@ import co.touchlab.kermit.Logger
 import com.github.submob.logmob.ANRWatchDogHandler
 import com.github.submob.logmob.initCrashlytics
 import com.github.submob.logmob.initLogger
-import com.oztechan.ccc.ad.initAds
 import com.oztechan.ccc.analytics.initAnalytics
 import com.oztechan.ccc.android.app.di.initKoin
+import com.oztechan.ccc.android.core.ad.initAds
 
 @Suppress("unused")
 class Application : Application() {
