+++ android/app/src/main/kotlin/com/oztechan/ccc/android/app/di/Koin.kt
@@ -3,10 +3,10 @@ package com.oztechan.ccc.android.app.di
 import android.content.Context
 import android.os.Build
 import co.touchlab.kermit.Logger
-import com.oztechan.ccc.ad.di.adModule
 import com.oztechan.ccc.analytics.di.analyticsModule
 import com.oztechan.ccc.android.app.BuildConfig
-import com.oztechan.ccc.billing.di.billingModule
+import com.oztechan.ccc.android.core.ad.di.androidCoreAdModule
+import com.oztechan.ccc.android.core.billing.di.androidCoreBillingModule
 import com.oztechan.ccc.client.core.persistence.di.clientCorePersistenceModule
 import com.oztechan.ccc.client.datasource.currency.di.clientDataSourceCurrencyModule
 import com.oztechan.ccc.client.datasource.watcher.di.clientDataSourceWatcherModule
@@ -31,14 +31,15 @@ fun initKoin(context: Context) = startKoin {
 
     modules(
         module { singleOf(::provideDevice) },
-        adModule,
         analyticsModule,
-        billingModule,
 
         viewModelModule,
         repositoryModule,
 
         configModule,
+
+        androidCoreAdModule,
+        androidCoreBillingModule,
 
         // client
         clientCorePersistenceModule,
