+++ android/core/ad/src/google/kotlin/com/oztechan/ccc/android/core/ad/Ads.kt
@@ -1,0 +1,10 @@
+package com.oztechan.ccc.android.core.ad
+
+import android.content.Context
+import co.touchlab.kermit.Logger
+import com.google.android.gms.ads.MobileAds
+
+fun initAds(context: Context) {
+    Logger.i { "Ads initAds" }
+    MobileAds.initialize(context)
+}
