+++ android/core/ad/src/huawei/kotlin/com/oztechan/ccc/android/core/ad/Ads.kt
@@ -1,0 +1,9 @@
+package com.oztechan.ccc.android.core.ad
+
+import android.content.Context
+import co.touchlab.kermit.Logger
+
+@Suppress("UNUSED_PARAMETER")
+fun initAds(context: Context) {
+    Logger.i { "Ads initAds" }
+}
