+++ android/core/ad/src/main/AndroidManifest.xml
