+++ android/core/ad/src/main/kotlin/com/oztechan/ccc/android/core/ad/AdManager.kt
@@ -1,0 +1,27 @@
+package com.oztechan.ccc.android.core.ad
+
+import android.app.Activity
+import android.content.Context
+
+interface AdManager {
+
+    fun getBannerAd(
+        context: Context,
+        width: Int,
+        adId: String,
+        onAdLoaded: (Int?) -> Unit
+    ): BannerAdView
+
+    fun showInterstitialAd(
+        activity: Activity,
+        adId: String
+    )
+
+    fun showRewardedAd(
+        activity: Activity,
+        adId: String,
+        onAdFailedToLoad: () -> Unit,
+        onAdLoaded: () -> Unit,
+        onReward: () -> Unit
+    )
+}
