+++ android/core/billing/src/main/kotlin/com/oztechan/ccc/android/core/billing/BillingManager.kt
@@ -1,0 +1,20 @@
+package com.oztechan.ccc.android.core.billing
+
+import android.app.Activity
+import androidx.lifecycle.LifecycleCoroutineScope
+import kotlinx.coroutines.flow.SharedFlow
+
+interface BillingManager {
+    val effect: SharedFlow<BillingEffect>
+
+    fun startConnection(
+        lifecycleScope: LifecycleCoroutineScope,
+        skuList: List<String>
+    )
+
+    fun endConnection()
+
+    fun launchBillingFlow(activity: Activity, skuId: String)
+
+    fun acknowledgePurchase()
+}
