+++ android/core/billing/src/main/kotlin/com/oztechan/ccc/android/core/billing/model/ProductDetails.kt
@@ -1,0 +1,7 @@
+package com.oztechan.ccc.android.core.billing.model
+
+data class ProductDetails(
+    val price: String,
+    val description: String,
+    val id: String
+)
