+++ android/core/billing/src/main/kotlin/com/oztechan/ccc/android/core/billing/model/PurchaseHistoryRecord.kt
@@ -1,0 +1,6 @@
+package com.oztechan.ccc.android.core.billing.model
+
+data class PurchaseHistoryRecord(
+    var ids: List<String>,
+    var date: Long
+)
