+++ android/feature/mobile/build.gradle.kts
@@ -77,10 +77,13 @@ dependencies {
 
     implementation(project(Modules.Client.self))
 
-    Modules.apply {
-        implementation(project(res))
+    Modules.Android.Core.apply {
         implementation(project(billing))
         implementation(project(ad))
+    }
+
+    Modules.apply {
+        implementation(project(res))
         implementation(project(analytics))
     }
