+++ android/feature/mobile/src/test/kotlin/com/oztechan/ccc/android/feature/mobile/util/PurchaseUtilTest.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.android.feature.mobile.util
 
-import com.oztechan.ccc.billing.model.ProductDetails
-import com.oztechan.ccc.billing.model.PurchaseHistoryRecord
+import com.oztechan.ccc.android.core.billing.model.ProductDetails
+import com.oztechan.ccc.android.core.billing.model.PurchaseHistoryRecord
 import com.oztechan.ccc.client.model.PremiumType
 import kotlin.test.Test
 import kotlin.test.assertEquals
