--- billing/src/main/kotlin/com/oztechan/ccc/billing/di/BillingModule.kt
@@ -1,11 +1,0 @@
-package com.oztechan.ccc.billing.di
-
-import com.oztechan.ccc.billing.BillingManager
-import com.oztechan.ccc.billing.BillingManagerImpl
-import org.koin.core.module.dsl.bind
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.module
-
-val billingModule = module {
-    singleOf(::BillingManagerImpl) { bind<BillingManager>() }
-}
