--- billing/src/main/kotlin/com/oztechan/ccc/billing/model/ProductDetails.kt
@@ -1,7 +1,0 @@
-package com.oztechan.ccc.billing.model
-
-data class ProductDetails(
-    val price: String,
-    val description: String,
-    val id: String
-)
