--- billing/src/main/kotlin/com/oztechan/ccc/billing/model/PurchaseHistoryRecord.kt
@@ -1,6 +1,0 @@
-package com.oztechan.ccc.billing.model
-
-data class PurchaseHistoryRecord(
-    var ids: List<String>,
-    var date: Long
-)
