+++ settings.gradle.kts
@@ -24,6 +24,8 @@ dependencyResolutionManagement {
 include(
     // Targets
     ":android:app", // android app
+    "android:core:billing", // android only billing module
+    "android:core:ad", // android only ad module
     ":android:feature:mobile", // android mobile app
     ":android:feature:widget", // android widget
 
@@ -56,9 +58,6 @@ include(
     ":analytics", // Shared with all FE targets for Google Analytics
 
     ":provider", // umbrella framework for iOS libraries
-
-    ":billing", // android only billing module
-    ":ad", // android only ad module
 
     // submodules
     ":submodule:logmob", // KMP, logger library
